move old policy decay to common utils
